--- src/freeFunctions/game/areMetroidsDead.cpp
@@ -1,17 +1,0 @@
-#include <vector>
-#include "../../../include/models/iMetroid.hpp"
-
-bool areMetroidsDead(const std::vector<std::vector<IMetroid *>> &metroids)
-{
-  for (auto row : metroids)
-  {
-    for (auto metroid : row)
-    {
-      if (metroid->isAlive())
-      {
-        return false;
-      }
-    }
-  }
-  return true;
-}
